CLRadeonExtender: Fixed ParseException constructors (use uint64_t instead size_t). Added reading of character literals in assembler expressions. Remove obsolete methods.

+++ CLRX/utils/Utilities.h
@@ -66,8 +66,8 @@ class ParseException: public Exception
 public:
     ParseException() = default;
     explicit ParseException(const std::string& message);
-    ParseException(size_t lineNo, const std::string& message);
-    ParseException(size_t lineNo, size_t charNo, const std::string& message);
+    ParseException(uint64_t lineNo, const std::string& message);
+    ParseException(uint64_t lineNo, size_t charNo, const std::string& message);
     virtual ~ParseException() throw() = default;
 };
